+++ CHANGES.md
@@ -12,6 +12,8 @@ E-Maj - Change log
   * Record the E-Maj rollback start, initialization end and locking end
     timestamps into the emaj_rlbk table and let the emaj_rollback_activity()
     function report both rollback planning and locking phases duration.
+  * Allow rollback duration estimate functions to be executed while a rollback
+    operation is in progress for the same tables groups.
   * Minor code changes.
 
 ###Bug fixes:###